+++ src/IkeMtz.NRSRx.Core.OData/CoreODataStartup.cs
@@ -6,14 +6,18 @@ using Microsoft.AspNet.OData.Builder;
 using Microsoft.AspNet.OData.Extensions;
 using Microsoft.AspNet.OData.Formatter;
 using Microsoft.AspNet.OData.Formatter.Serialization;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.ApiExplorer;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Options;
 using Microsoft.Net.Http.Headers;
 using Microsoft.OData;
+using Microsoft.OpenApi.Models;
+using Swashbuckle.AspNetCore.SwaggerGen;
 
 namespace IkeMtz.NRSRx.Core.OData
 {
@@ -116,6 +120,51 @@ namespace IkeMtz.NRSRx.Core.OData
             options.SubstituteApiVersionInUrl = true;
           });
       return mvcBuilder;
+    }
+
+    public void SetupSwagger(IServiceCollection services)
+    {
+      _ = services
+        .AddTransient<IConfigureOptions<SwaggerGenOptions>>(serviceProvider => new ConfigureSwaggerOptions(serviceProvider.GetRequiredService<IApiVersionDescriptionProvider>(), this))
+        .AddSwaggerGen(options =>
+        {
+          options.OperationFilter<ODataCommonOperationFilter>();
+          options.UseInlineDefinitionsForEnums();
+          // add a custom operation filter which sets default values
+          options.OperationFilter<SwaggerDefaultValues>();
+          var audiences = GetIdentityAudiences();
+          var swaggerIdentityProviderUrl = Configuration.GetValue<string>("SwaggerIdentityProviderUrl");
+          if (audiences.Any() && !string.IsNullOrWhiteSpace(swaggerIdentityProviderUrl))
+          {
+            var audience = audiences.FirstOrDefault();
+
+            options.AddSecurityDefinition(JwtBearerDefaults.AuthenticationScheme, new OpenApiSecurityScheme
+            {
+              Type = SecuritySchemeType.OAuth2,
+              In = ParameterLocation.Header,
+              Description = "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
+              Scheme = JwtBearerDefaults.AuthenticationScheme,
+              Flows = new OpenApiOAuthFlows
+              {
+                Implicit = new OpenApiOAuthFlow
+                {
+                  AuthorizationUrl = new Uri($"{swaggerIdentityProviderUrl}authorize?audience={audience}"),
+                  Scopes = SwaggerScopes,
+                },
+              }
+            });
+            options.AddSecurityRequirement(new OpenApiSecurityRequirement
+                {
+                    {
+                        new OpenApiSecurityScheme
+                        {
+                            Reference = new OpenApiReference { Type = ReferenceType.SecurityScheme, Id = JwtBearerDefaults.AuthenticationScheme}
+                        },
+                       Array.Empty<string>()
+                    }
+                });
+          }
+        });
     }
   }
 }
