+++ src/IkeMtz.NRSRx.Core.OData/GlobalSuppressions.cs
@@ -1,0 +1,8 @@
+// This file is used by Code Analysis to maintain SuppressMessage
+// attributes that are applied to this project.
+// Project-level suppressions either have no target or are given
+// a specific target and scoped to a namespace, type, member, etc.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Design", "CA1062:Validate arguments of public methods", Justification = "<Pending>", Scope = "member", Target = "~M:IkeMtz.NRSRx.Core.OData.ODataCommonOperationFilter.AddParameter(Microsoft.OpenApi.Models.OpenApiOperation,System.String,System.String)")]
