+++ src/IkeMtz.NRSRx.Core.Web/SwaggerDefaultValues.cs
@@ -27,9 +27,12 @@ namespace IkeMtz.NRSRx.Core.Web
       // REF: https://github.com/domaindrivendev/Swashbuckle.AspNetCore/pull/413
       foreach (var parameter in operation.Parameters)
       {
-        var description = apiDescription.ParameterDescriptions.First(p => p.Name == parameter.Name);
-
-        if (parameter.Description == null)
+        var description = apiDescription.ParameterDescriptions.FirstOrDefault(p => p.Name == parameter.Name);
+        if (description == null)
+        {
+          continue;
+        }
+        else if (parameter.Description == null)
         {
           parameter.Description = description.ModelMetadata?.Description;
         }
